requirements.txt
tensorflow>=2.10.0
pandas
pyyaml
